part of #567: move xml_2_json.h/cc to util
